Argument Parser Issue Fix
Issue: We noticed that during training,
the model was overwriting the "freeze" == "True" version
of the model even when the "freeze" == "False" version was passed.

How:The argument parser cannot read true and false values directly
from the CLI, which means in the old version, whether we pass "True" or "False"
the parser will always interpret it as True since it is a string.

Fix: To fix it, we change how the parameter works. If the parameter is passed,
it will be interpreted as True, otherwise it will be False.