Argument Frozen gradient Computation
Issue: After fixing the argument parser bug, we noticed that "Freeze" was hardcoded to False
in another part of the code. When we tried to pass "Freeze" as True, the code
would error out saying "element 0 of variables does not require grad and does not have a grad_fn".
This indicates that freezing the base model never worked before, we just assumed it was working.

How: The reason this issue happens is because first, we tried to optimize
supposedly frozen gradients, which is not possible. Second, the forward
pass always returns the loss from the base model. but, if the base model
is frozen, the loss will not have requires_grad=True, so the optimizer
will not be able to optimize it.

Fix: First, we update the optimizer to only optimize the parameters that
have requires_grad=True. Second, we update the forward pass to set
requires_grad=True for the loss. This way, the optimizer can optimize
the loss. Note that even though the base model is frozen,the gradients
still exist for those layers, but the optimizer will not
update them. That is why we can still optimize the loss and update the
preprocess layer via it. Otherwise this solution would not work since
setting requires_grad=True for the loss would not magically compute
gradients for the frozen layers at the end of a forward pass.